+++ src/config.rs
@@ -3,7 +3,10 @@ use std::collections::BTreeMap;
 use anyhow::Result;
 use serde::Deserialize;
 
-use crate::{cli::CliArgs, nix};
+use crate::{
+    cli::CliArgs,
+    nix::{self, url::FlakeUrl},
+};
 
 /// Rust type for the `nixci` flake output
 ///
@@ -30,8 +33,11 @@ impl Default for Config {
 }
 
 impl Config {
-    pub fn from_flake_url(url: &str) -> Result<Self> {
-        nix::eval::nix_eval_attr_json::<Config>("nixci", url)
+    pub fn from_flake_url(url: &FlakeUrl) -> Result<(Self, FlakeUrl)> {
+        let (url, attr) = url.split_attr();
+        let nixci_url = FlakeUrl(format!("{}#nixci.{}", url.0, attr.get_name()));
+        let cfg = nix::eval::nix_eval_attr_json::<Config>(nixci_url)?;
+        Ok((cfg, url))
     }
 }
 
@@ -48,7 +54,7 @@ pub struct SubFlakish {
     // NB: we use BTreeMap instead of HashMap here so that we always iterate
     // inputs in a determinitstic (i.e. asciibetical) order
     #[serde(rename = "overrideInputs", default)]
-    pub override_inputs: BTreeMap<String, String>,
+    pub override_inputs: BTreeMap<String, FlakeUrl>,
 }
 
 impl Default for SubFlakish {
@@ -62,19 +68,14 @@ impl Default for SubFlakish {
 }
 
 impl SubFlakish {
-    /// Return the flake URL pointing to the sub-flake
-    pub fn sub_flake_url(&self, root_flake_url: &String) -> String {
-        if self.dir == "." {
-            root_flake_url.clone()
-        } else {
-            format!("{}?dir={}", root_flake_url, self.dir)
-        }
-    }
-
     /// Return the `nix build` arguments for building all the outputs in this
     /// subflake configuration.
-    pub fn nix_build_args_for_flake(&self, cli_args: &CliArgs, flake_url: &String) -> Vec<String> {
-        std::iter::once(self.sub_flake_url(flake_url))
+    pub fn nix_build_args_for_flake(
+        &self,
+        cli_args: &CliArgs,
+        flake_url: &FlakeUrl,
+    ) -> Vec<String> {
+        std::iter::once(flake_url.sub_flake_url(self.dir.clone()).0)
             .chain(self.override_inputs.iter().flat_map(|(k, v)| {
                 [
                     "--override-input".to_string(),
@@ -82,7 +83,7 @@ impl SubFlakish {
                     // devour-flake uses that input name to refer to the user's
                     // flake.
                     format!("flake/{}", k),
-                    v.to_string(),
+                    v.0.to_string(),
                 ]
             }))
             .chain(cli_args.extra_nix_build_args.iter().cloned())