+++ src/nix/lock.rs
@@ -4,14 +4,13 @@ use anyhow::{bail, Result};
 
 use crate::nix::util::print_shell_command;
 
-pub fn nix_flake_lock_check(url: &str) -> Result<()> {
-    print_shell_command(
-        "nix",
-        ["flake", "lock", "--no-update-lock-file", url].into_iter(),
-    );
+use super::url::FlakeUrl;
+
+pub fn nix_flake_lock_check(url: &FlakeUrl) -> Result<()> {
+    let args = ["flake", "lock", "--no-update-lock-file", &url.0];
+    print_shell_command("nix", args.into_iter());
     let status = Command::new("nix")
-        .args(["flake", "lock", "--no-update-lock-file"])
-        .arg(url)
+        .args(args)
         .stdin(Stdio::null())
         .spawn()?
         .wait()?;
@@ -19,6 +18,6 @@ pub fn nix_flake_lock_check(url: &str) -> Result<()> {
         Ok(())
     } else {
         let exit_code = status.code().unwrap_or(1);
-        bail!("nix eval failed to run (exited: {})", exit_code);
+        bail!("nix flake lock failed to run (exited: {})", exit_code);
     }
 }